Modifies getPollBooks.php API
This commit updates getPollBooks.php with a few enhancements.
Now uses prepared statements for queries to avoid any issues with
SQL injection; enables a search by constituency (via a join with
the elections table); enables inclusion of election results as an
option. Simplified query-building code a bit.
